Refactor ProgramRepository
