+++ app/src/main/java/org/eyeseetea/malariacare/domain/boundary/repositories/IProgramRepository.java
@@ -7,7 +7,5 @@ public interface IProgramRepository {
 
     void saveUserProgramId(Program program);
 
-    Long getUserProgramId();
-
-    String getUserProgramUID();
+    Program getUserProgram();
 }
